moved things in attribute_combining_functions.py to combining main. moved definitions to scraping folder
